requirements.txt
Django==1.9.7
django-debug-toolbar==1.4
djangorestframework==3.3.3
psycopg2==2.6.1
sqlparse==0.1.19
